fix(gluon): emit lowercase cfg names to match code conventions
Gluon's Kconfig option emission used the raw option name (e.g. `ACPI`)
producing `--cfg hadron_ACPI`, but all kernel code uses lowercase
`#[cfg(hadron_acpi)]`. This made all cfg-gated code dead since the
project's inception.

### Fixed
- Lowercase Kconfig option names before emitting `--cfg hadron_<name>`
- Fix pre-existing dead code exposed by correct cfg activation:
  - gdt.rs: `crate::mm::vmm::with` → `crate::vmm::with`
  - hpet.rs: remove non-existent `ClockSource` import and impl
  - idt.rs: remove non-existent `timer_preempt_stub` usage
  - smp.rs: `#[no_mangle]` → `#[unsafe(no_mangle)]`, fix hhdm type
  - percpu.rs: fix `get_for()` argument type (usize → u32)

+++ kernel/kernel/src/percpu.rs
@@ -109,7 +109,7 @@ pub fn init_ap_percpu(cpu_id: u32, kernel_rsp: u64) -> *mut PerCpuState {
         _pad2: [0; 2],
         _pad3: [0; 24],
         saved_regs_ptr: crate::arch::x86_64::syscall::SYSCALL_SAVED_REGS
-            .get_for(cpu_id as usize)
+            .get_for(cpu_id)
             .get() as u64,
     });
     let ptr = Box::into_raw(state);
